requirements.txt
gradio>=4.0.0
huggingface_hub>=0.20.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-pinecone>=0.1.0
pinecone-client>=3.0.0
requests>=2.28.0
